% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediccion.R
\name{prediccion}
\alias{prediccion}
\title{Recibe un registro en JSON y devuleve una predicción del mismo en json}
\usage{
prediccion(registro)
}
\arguments{
\item{registro}{}
}
\value{
json
}
\description{
Recibe un registro en JSON y devuleve una predicción del mismo en json
}
\examples{
 prediccion(5)

}
