% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prediccion.R
\name{prediccion}
\alias{prediccion}
\title{Recibe un registro en JSON y devuleve una predicción del mismo en json}
\usage{
prediccion(
  ContadorBajasCCC,
  anyconti,
  numconti,
  ContadordiasBajasDNI,
  sexo,
  cnae,
  fecha.baja,
  fecha.alta,
  icd9,
  recaida,
  num.trabajadores.emp,
  codigo.postal,
  fecha.alta.empresa,
  contrato,
  grupo.cotizacion,
  pluriempleo,
  diasemana,
  esprincipal,
  fecha.nacimiento
)
}
\arguments{
\item{ContadorBajasCCC, }{anyconti, numconti, ContadordiasBajasDNI, sexo, cnae, fecha.baja, fecha.alta, icd9, recaida, num.trabajadores.emp, codigo.postal, fecha.alta.empresa, contrato, grupo.cotizacion, pluriempleo, diasemana, esprincipal, fecha.nacimiento}
}
\value{
json
}
\description{
Recibe un registro en JSON y devuleve una predicción del mismo en json
}
\examples{
 prediccion("8","2010","29754A","78","1","8122","15/12/2016","-","727.61","","7","25600","12/4/1999","100","8","","jue.","1","10/5/1958")

}
